pyproject.toml
[tool.poetry]
name = "deepgreat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
mysqlclient = "^2.2.4"
dotenv-python = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
